fix: Fix test compilation errors and improve test patterns
- Rewrite contextual_typing_tests.rs and spread_rest_tests.rs to use
  working check_source() pattern instead of broken create_checker()
  that caused borrow checker issues (E0515/E0505)
- Fix u16/u32 type mismatch in control_flow.rs tests (cast flags as u32)
- Add missing 'target' field to CheckerOptions in checker_state_tests.rs
- Gate SalsaDatabase tests behind experimental_salsa feature in db_tests.rs
- Fix unstable str_as_str feature usage in template_expansion_tests.rs
- Update stability_validation_tests.rs to use public module paths
- Make MAX_INSTANTIATION_DEPTH, MAX_LOWERING_OPERATIONS,
  MAX_CONSTRAINT_RECURSION_DEPTH public for tests
- Fix SubtypeFailureReason import path in compat_tests.rs
- Update test_fixtures.rs to use checker::context::LibContext

+++ src/checker/control_flow.rs
@@ -3765,7 +3765,7 @@ if (typeof x === "string") {
         let var_decl_node = arena.get(var_decl_idx).expect("var decl node");
 
         // Verify the variable is let (not const)
-        let flags = var_decl_node.flags;
+        let flags = var_decl_node.flags as u32;
         let is_const = (flags & node_flags::CONST) != 0;
         assert!(!is_const, "Variable should be let (mutable), not const");
 
@@ -3813,7 +3813,7 @@ if (typeof x === "string") {
         let var_decl_node = arena.get(var_decl_idx).expect("var decl node");
 
         // Verify the variable is const
-        let flags = var_decl_node.flags;
+        let flags = var_decl_node.flags as u32;
         let is_const = (flags & node_flags::CONST) != 0;
         assert!(is_const, "Variable should be const");
 
@@ -3849,7 +3849,7 @@ const fn = () => {
         let var_decl_node = arena.get(var_decl_idx).expect("var decl node");
 
         // Verify the variable is let (mutable)
-        let flags = var_decl_node.flags;
+        let flags = var_decl_node.flags as u32;
         let is_const = (flags & node_flags::CONST) != 0;
         assert!(!is_const, "Variable should be let (mutable)");
     }
